move text fixtures into separate directories
