+++ fixtures/curl_commands/curl10.txt
