+++ fixtures/curl_commands/curl18.txt
