+++ fixtures/curl_commands/curl6.txt
