+++ fixtures/curl_commands/curl7.txt
