+++ fixtures/curl_commands/curl8.txt
