+++ fixtures/curl_commands/curl9.txt
