+++ fixtures/node_output/node_output1.js
