+++ fixtures/node_output/node_output10.js
