+++ fixtures/node_output/node_output11.js
