+++ fixtures/node_output/node_output12.js
