+++ fixtures/node_output/node_output13.js
