+++ fixtures/node_output/node_output2.js
