+++ fixtures/node_output/node_output3.js
