+++ fixtures/node_output/node_output4.js
