+++ fixtures/node_output/node_output5.js
