+++ fixtures/node_output/node_output7.js
