+++ fixtures/node_output/node_output8.js
