+++ fixtures/node_output/node_output9.js
