+++ fixtures/parser_output/parser_output2.js
