+++ fixtures/parser_output/parser_output3.js
