+++ fixtures/python_output/python_output10.py
