+++ fixtures/python_output/python_output3.py
