+++ fixtures/python_output/python_output6.py
