+++ fixtures/python_output/python_output8.py
