+++ fixtures/python_output/python_output9.py
